Mais edicoes na tela de cliente

+++ Academia/src/br/senai/sp/jandira/dao/ClienteDAO.java
@@ -31,7 +31,7 @@ public class ClienteDAO {
 			stm.setString(1, cliente.getNome());
 			stm.setString(2, cliente.getDtNasc());
 			stm.setInt(3, cliente.getPeso());
-			stm.setDouble(4, cliente.getAltura());
+			stm.setInt(4, cliente.getAltura());
 			stm.setString(5, cliente.getSexo());
 			stm.setString(6, cliente.getNivelAtividade());
 			stm.execute();
@@ -59,7 +59,7 @@ public class ClienteDAO {
 			stm.setString(1, cliente.getNome());
 			stm.setString(2, cliente.getDtNasc());
 			stm.setInt(3, cliente.getPeso());
-			stm.setDouble(4, cliente.getAltura());
+			stm.setInt(4, cliente.getAltura());
 			stm.setString(5, cliente.getSexo());
 			stm.setString(6, cliente.getNivelAtividade());
 			stm.setInt(7, cliente.getId());
@@ -115,7 +115,7 @@ public class ClienteDAO {
 			cliente.setNome(rs.getString("nome"));
 			cliente.setDtNasc(df.format(rs.getDate("dtNasc")));
 			cliente.setPeso(rs.getInt("peso"));
-			cliente.setAltura(rs.getDouble("altura"));
+			cliente.setAltura(rs.getInt("altura"));
 			cliente.setSexo(rs.getString("sexo"));
 			cliente.setNivelAtividade(rs.getString("niveldeatividade"));
 			
@@ -150,7 +150,7 @@ public class ClienteDAO {
 				cliente.setNome(rs.getString("nome"));
 				cliente.setDtNasc(rs.getString("dtNasc"));
 				cliente.setPeso(rs.getInt("peso"));
-				cliente.setAltura(rs.getDouble("altura"));
+				cliente.setAltura(rs.getInt("altura"));
 				cliente.setSexo(rs.getString("sexo"));
 				cliente.setNivelAtividade(rs.getString("niveldeatividade"));
 				clientes.add(cliente);
